added condiditon to migration file

+++ db/migrate/20250601180734_remove_year_option_from_contact.rb
@@ -1,6 +1,6 @@
 class RemoveYearOptionFromContact < ActiveRecord::Migration[8.0]
   def change
-    remove_column :contacts, :year_entered, :integer
-    remove_column :contacts, :year_left, :integer
+    remove_column :contacts, :year_entered, :integer if column_exists?(:contacts, :year_entered)
+    remove_column :contacts, :year_left, :integer if column_exists?(:contacts, :year_left)
   end
 end
